# openapi.yaml
openapi: 3.1.0
info:
  title: PDF Accessibility Platform API
  description: |
    API for PDF accessibility processing platform

    This API provides document processing capabilities to convert PDFs into accessible formats
    including ADA/Section 508 compliant PDFs and LLM-ready structured data exports.

    ## Authentication
    - Bearer token authentication using BetterAuth JWT tokens
    - API keys for programmatic access (coming soon)

    ## Rate Limits
    - 60 requests per minute per user
    - 1000 requests per hour per user

    ## Supported Formats
    - Input: PDF, DOC, DOCX, TXT files
    - Output: Accessible PDF, HTML, JSON, CSV/ZIP exports

  version: 1.0.0
  contact:
    name: PDF Accessibility Platform
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.pdf-accessibility.com
    description: Production server
...
paths:
...
  /reports/summary:
    get:
      tags: [reports]
      summary: Get reports summary
      description: Get summary statistics including document counts, averages by week, and success rates (admin only)
      operationId: get_summary_report
      security:
        - BetterAuthJWTBearer: []
      responses:
        '200':
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /reports/export.csv:
    get:
      tags: [reports]
      summary: Export documents to CSV
      description: Stream CSV export of documents with filtering support (admin only)
      operationId: export_documents_csv
      security:
        - BetterAuthJWTBearer: []
      parameters:
        - name: start_date
          in: query
          description: Start date filter (ISO format)
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: End date filter (ISO format)
          schema:
            type: string
            format: date-time
        - name: owner_filter
          in: query
          description: Filter by owner ID
          schema:
            type: string
        - name: status_filter
          in: query
          description: Filter by document status
          schema:
            type: string
      responses:
        '200':
          description: CSV file stream
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Webhooks
  /webhooks:
    post:
      tags: [webhooks]
      summary: Receive webhook callbacks
      description: Endpoint for receiving HMAC-signed webhook callbacks from external services
      operationId: receive_webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook received successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /webhooks/test:
    post:
      tags: [webhooks]
      summary: Test webhook endpoint
      description: Test endpoint for webhook functionality (development only)
      operationId: test_webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook test successful
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  securitySchemes:
    BetterAuthJWTBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from BetterAuth authentication
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ReportsSummaryResponse:
      type: object
      required:
        [
          total_documents,
          completed_documents,
          processing_documents,
          failed_documents,
          pending_documents,
          completion_rate,
          avg_processing_time_hours,
          weekly_stats,
        ]
      properties:
        total_documents:
          type: integer
          description: Total documents processed
        completed_documents:
          type: integer
          description: Successfully completed documents
        processing_documents:
          type: integer
          description: Documents currently being processed
        failed_documents:
          type: integer
          description: Documents that failed processing
        pending_documents:
          type: integer
          description: Documents pending processing
        completion_rate:
          type: number
          description: Success rate percentage (0-100)
        avg_processing_time_hours:
          type: number
          description: Average processing time in hours
        weekly_stats:
          type: array
          items:
            type: object
            additionalProperties: true
          description: Weekly statistics with aggregated data

    # Error responses
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...